Move files around (#511)
* Mac finder metadata to .gitignore

* Remove data module, move content to ureg/ and isotopes/

This gives a generally cleaner structure: data files live in /data/
subdirectories of their relevant packages.

While we're at it, split content of isotopes.py to isotopes/_core.py and isotopes/_csv.py and import useful things to isotope/__init__.py. Keeps the files a bit more navigable without complicating public API too much.

Annoyingly it does bring back some of the import-loop headache for
legacy data loader; isotopes/_legacy.py can't be loaded independently
of isotopes/__init__.py which creates a circular dependence on
euphonic.util.  The long-term fix is clear (drop isotopes._legacy) and
the real sin was having functionality of util depend on another module.

* Update CHANGELOG
