+++ euphonic/isotopes/data/sears-1992.json
