+++ euphonic/ureg/data/reciprocal_spectroscopy_definitions.txt
